+++ EnergyPerformance/Services/Controller.cs
@@ -1,0 +1,29 @@
+﻿namespace EnergyPerformance.Services;
+
+public class Controller
+{
+    public int PerformanceCoreCount()
+    {
+        return 4;
+    }
+
+    public int EfficiencyCoreCount() 
+    {
+        return 4;
+    }
+
+    public int TotalCoreCount()
+    {
+        return 12;
+    }
+
+    public void MoveAllAppsToHybridCores()
+    {
+        return;
+    }
+
+    public void ResetToDefaultCores()
+    {
+        return;
+    }
+}
