+++ EnergyPerformance/app.manifest
@@ -6,14 +6,6 @@
         <supportedOS Id="{8e0f7a12-bfb3-4fe8-b9a5-48fd50a15a9a}"/>
       </application>
     </compatibility>
-	<!--comment out the entire <trustInfo> section below to disable request for admin priveleges on startup-->
-    <trustInfo xmlns="urn:schemas-microsoft-com:asm.v3">
-      <security>
-        <requestedPrivileges>
-			<requestedExecutionLevel level="requireAdministrator" uiAccess="false"/>
-        </requestedPrivileges>
-      </security>
-    </trustInfo>
   <application xmlns="urn:schemas-microsoft-com:asm.v3">
     <windowsSettings>
       <dpiAware xmlns="http://schemas.microsoft.com/SMI/2005/WindowsSettings">true/PM</dpiAware>
